+++ CI/installation/install_packages_fedora:38.sh
@@ -5,7 +5,7 @@ set -ex
 # Install packages
 sudo dnf install -y gcc g++ glibc-static libstdc++-static git make cmake autoconf \
     diffutils findutils wget which procps squashfs-tools zlib-devel zlib-static \
-    boost-devel skopeo tini-static \
+    boost-devel skopeo runc tini-static \
     python3 python3-pip python3-setuptools
 sudo dnf clean all
 sudo rm -rf /var/cache/dnf