+++ CI/src/integration_tests/test_glibc_hook.py
@@ -113,9 +113,9 @@ class TestGlibcHook(unittest.TestCase):
     def _pull_docker_images(cls):
         util.pull_image_if_necessary(is_centralized_repository=False, image=util.ALPINE_IMAGE)            # no glibc
         util.pull_image_if_necessary(is_centralized_repository=False, image="quay.io/ethcscs/centos:6")   # glibc 2.12
-        util.pull_image_if_necessary(is_centralized_repository=False, image="quay.io/ethcscs/fedora:36")  # assumption: glibc >= host's glibc
+        util.pull_image_if_necessary(is_centralized_repository=False, image="quay.io/ethcscs/fedora:38")  # assumption: glibc >= host's glibc
         # based on fedora - assumption: glibc >= host's glibc
-        util.pull_image_if_necessary(is_centralized_repository=False, image="quay.io/ethcscs/sarus-integration-tests:nonexisting_ldcache_entry_f36")
+        util.pull_image_if_necessary(is_centralized_repository=False, image="quay.io/ethcscs/sarus-integration-tests:nonexisting_ldcache_entry_f38")
 
     @classmethod
     def _enable_hook(cls):
@@ -155,13 +155,13 @@ class TestGlibcHook(unittest.TestCase):
 
     def test_no_injection_in_container_with_recent_glibc(self):
         self._glibc_command_line_option = True
-        self._container_image = "quay.io/ethcscs/fedora:36"
+        self._container_image = "quay.io/ethcscs/fedora:38"
         hashes = self._get_hashes_of_host_libs_in_container()
         assert not hashes
 
     def test_no_injection_in_container_with_recent_glibc_and_nonexisting_ldcache_entry(self):
         self._glibc_command_line_option = True
-        self._container_image = "quay.io/ethcscs/sarus-integration-tests:nonexisting_ldcache_entry_f36"
+        self._container_image = "quay.io/ethcscs/sarus-integration-tests:nonexisting_ldcache_entry_f38"
         hashes = self._get_hashes_of_host_libs_in_container()
         assert not hashes